+++ modules/api/api.py
@@ -46,6 +46,7 @@ class Api:
         self.add_api_route("/sdapi/v1/status", server.get_status, methods=["GET"], response_model=models.ResStatus)
         self.add_api_route("/sdapi/v1/platform", server.get_platform, methods=["GET"])
         self.add_api_route("/sdapi/v1/progress", server.get_progress, methods=["GET"], response_model=models.ResProgress)
+        self.add_api_route("/sdapi/v1/history", server.get_history, methods=["GET"], response_model=list[models.ResHistory])
         self.add_api_route("/sdapi/v1/interrupt", server.post_interrupt, methods=["POST"])
         self.add_api_route("/sdapi/v1/skip", server.post_skip, methods=["POST"])
         self.add_api_route("/sdapi/v1/shutdown", server.post_shutdown, methods=["POST"])
@@ -92,8 +93,8 @@ class Api:
         self.add_api_route("/sdapi/v1/unload-checkpoint", endpoints.post_unload_checkpoint, methods=["POST"])
         self.add_api_route("/sdapi/v1/reload-checkpoint", endpoints.post_reload_checkpoint, methods=["POST"])
         self.add_api_route("/sdapi/v1/refresh-vae", endpoints.post_refresh_vae, methods=["POST"])
-        self.add_api_route("/sdapi/v1/history", endpoints.get_history, methods=["GET"], response_model=List[str])
-        self.add_api_route("/sdapi/v1/history", endpoints.post_history, methods=["POST"], response_model=int)
+        self.add_api_route("/sdapi/v1/latents", endpoints.get_latent_history, methods=["GET"], response_model=List[str])
+        self.add_api_route("/sdapi/v1/latents", endpoints.post_latent_history, methods=["POST"], response_model=int)
 
         # lora api
         if shared.native: